improving the classes

+++ src/game/characters/stats.py
@@ -1,5 +1,7 @@
 class Stats:
-  def __init__(self, move_speed: int = 1, max_hp: int = 1, attack_speed: int = 1) -> None:
+  def __init__(self, move_speed: float = 1, max_hp: int = 1, 
+               attack_speed: float = 1) -> None:
     self.move_speed = move_speed
     self.max_hp = max_hp
     self.attack_speed = attack_speed
+
